09-autonomous-flow/backend/requirements.txt
# AutonomousFlow - Dynamic Workflow Discovery
# Requirements for tool discovery and workflow regeneration

# Core dependencies
pyyaml>=6.0
pydantic>=2.5.0
httpx>=0.27.0

# Universal Agent Nexus (for compilation)
universal-agent-nexus>=1.0.3

# LangGraph (for runtime)
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-ollama>=0.2.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

